backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.18

# Validation
pydantic==2.10.5
pydantic-settings==2.7.0
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1

# Rate limiting (optional)
slowapi==0.1.9
